Fix non-native compilation
