+++ crates/module-system/sov-state/Cargo.toml
@@ -30,7 +30,7 @@ hex = { workspace = true }
 jmt = { workspace = true, features = ["std", "sha2"] }
 nomt-core = { workspace = true }
 nomt = { workspace = true, optional = true }
-sov-metrics = { workspace = true, optional = true }
+sov-metrics = { workspace = true }
 sov-modules-macros = { workspace = true, optional = true }
 tracing = { workspace = true, optional = true }
 
@@ -48,7 +48,7 @@ arbitrary = [
     "sov-rollup-interface/arbitrary",
     "sov-modules-macros?/arbitrary"
 ]
-bench = ["sov-metrics", "sov-modules-macros/bench"]
+bench = ["sov-modules-macros/bench"]
 default = []
 native = [
     "sov-db",
